Speeds up ``is_successful`` function
Here are the numbers:

  In [11]: %timeit for i in range(10000): f1(r(i))
  32.7 ms ± 721 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

  In [12]: %timeit for i in range(10000): f2(r(i))
  27.8 ms ± 191 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

Where `f1` is the old implementation, `f2` is the new one.
And `r` is a function to return both successful and failed values:

  r = lambda i: Success(1) if i % 2 else Failure(1)

I guess that this might be a hot path, so it is pretty good.
We also changed the input type.
Now we allow anything with the `.success_type` field.