fix (laws/pure): compatibility with hypothesis > 6.113.0 (#1956)
Hypothesis is now setting two entries for the Callables strategies, and then it
prefers the one using the type origin. While returns lookup for the first set
and replaces the strategy, starting with the one without the type origin.

This only changes the callable strategy search order to match hypothesis,
alternatively we could set both.