laws: restrict interface-patching to interfaces with laws.
We check that the interface class actually has laws. This should
eliminate any false positives that come from patching interfaces that
were defined outside `returns`.

Added a `_ParentWrapper` to test that it doesn't show up. (It did before
the change in this commit.)

Add to CHANGELOG and the hypothesis plugins page.