Make the `hypothesis` plugin check laws from user-defined interfaces too (#2060)
* laws: add xfail-ing test for custom interface with laws.

When running property tests for an container, we register it as the
strategy for each ancestor interface with laws.

However, we do that only for modules in `returns.`, which means that
lawful interfaces written by users will not be resolved. This commit
just demonstrates the current state.

* laws: extract and test function for getting lawful interfaces.

Right now, it just gets all MRO classes, and it misses interfaces
defined outside `returns`. Will tackle that in the next commit.

* laws: restrict interface-patching to lawful interfaces.

Earlier, we were restricting it to any class within `returns`.

Main user-visible change: We now include `Lawful` ancestors outside `returns`.

* laws: restrict interface-patching to interfaces with laws.

We check that the interface class actually has laws. This should
eliminate any false positives that come from patching interfaces that
were defined outside `returns`.

Added a `_ParentWrapper` to test that it doesn't show up. (It did before
the change in this commit.)

Add to CHANGELOG and the hypothesis plugins page.