hypothesis: add post-init check; make settings private.
Ideally, we would enforce the mutually exclusive settings in the types,
but I'm leaving it at this for now.

I'm making settings private again, because `check_all_laws` accepts
direct arguments, not a settings object. So, I don't think it makes much
sense to document the class in public and make it harder to change in
the future. I'd made it public just so that I could use it in tests.